Refactored some tests files
